# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Curd APIs
# during dev, should point to your local machine
host: localhost:32008
...
basePath: /user
...
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /googleauth:
    # binds a127 app logic to a route
    x-swagger-router-controller: google
    post:
      description: Auth google sign-in
      # used as the method name of the controller
      operationId: googleAuth
      parameters:
        - name: data
          in: body
          description: authenticate the user
          schema:
            $ref: '#/definitions/Request'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Request:
    type: object
  SuccessResponse:
    type: object
  ErrorResponse:
      type: object